cloud_orchestrator/fastapi_webserver/requirements.txt
uvicorn==0.12.2
fastapi==0.65.1
fastapi-versioning==0.9.1
SQLAlchemy==1.4.14
dependency-injector==4.32.2
python-jenkins==1.7.0
xmltodict==0.12.0
pylint==2.9.3
pylint-json2html
SQLAlchemy-Paginator
psycopg2-binary
pika
voluptuous
jsonschema
pytz
PyYAML==5.4.1
ldap3==2.9.1
python-gflags==3.1.2
cryptography==3.4.7
pycrypto==2.6.1
pandas
azure-devops
pycryptodome==3.10.1
adal
itsdangerous
authlib
httpx
prometheus-client==0.13.1
opentelemetry-instrumentation-fastapi
opentelemetry-exporter.prometheus
opentelemetry.exporter.otlp
opentelemetry.instrumentation.logging
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
prometheus-fastapi-instrumentator
